+++ src/components/DarkMode/DarkMode.html
@@ -1,29 +1,29 @@
-<div class="dark-mode">
+<div class="dark-mode-selector">
   <div class="mdc-menu-surface--anchor">
-    <div class="mdc-menu-surface dark-mode__holder" tabindex="-1">
+    <div class="mdc-menu-surface dark-mode-selector__holder" tabindex="-1">
       <ul
-        class="mdc-list dark-mode__list"
+        class="mdc-list dark-mode-selector__list"
         role="menu"
         aria-hidden="true"
         aria-orientation="vertical"
       >
         <li
-          class="mdc-list-item mdc-typography--body1 dark-mode__item dark-mode__item_dark"
+          class="mdc-list-item mdc-typography--body1 dark-mode-selector__item dark-mode-selector__item_dark"
           role="menuitem"
         >
           <span
-            class="mdc-list-item__text material-icons dark-mode__icon dark-mode__icon_black"
+            class="mdc-list-item__text material-icons dark-mode-selector__icon dark-mode-selector__icon_black"
           >
             brightness_1
           </span>
         </li>
 
         <li
-          class="mdc-list-item mdc-typography--body1 dark-mode__item dark-mode__item_light"
+          class="mdc-list-item mdc-typography--body1 dark-mode-selector__item dark-mode-selector__item_light"
           role="menuitem"
         >
           <span
-            class="mdc-list-item__text material-icons dark-mode__icon dark-mode__icon_purple"
+            class="mdc-list-item__text material-icons dark-mode-selector__icon dark-mode-selector__icon_purple"
           >
             brightness_1
           </span>
@@ -31,7 +31,7 @@
       </ul>
     </div>
   </div>
-  <button class="mdc-icon-button material-icons dark-mode__button">
+  <button class="mdc-icon-button material-icons dark-mode-selector__button">
     format_color_fill
   </button>
 </div>
