+++ src/components/DarkMode/DarkMode.scss
@@ -1,4 +1,4 @@
-.dark-mode {
+.dark-mode-selector {
   &__holder {
     margin-top: 3rem;
   }
