+++ src/components/Main/Main.js
@@ -1,6 +1,5 @@
 import { MDCDrawer } from "@material/drawer";
 import { AuthService } from "../../services/AuthService";
-import { MusicService } from "../../services/MusicService";
 import { MediaPlayerComponent } from "../MediaPlayer/MediaPlayer";
 import { HeaderComponent } from "../Header/Header";
 import { DarkModeComponent } from "../DarkMode/DarkMode";
@@ -80,14 +79,6 @@ export class MainComponent {
       "click",
       this.handleListClick.bind(this)
     );
-  }
-
-  async checkTheme() {
-    const user = await AuthService.check();
-    const currentTheme = await MusicService.getTheme(user.uid);
-    if (currentTheme === "dark") {
-      document.documentElement.classList.add("page_dark-mode");
-    }
   }
 
   handleListClick(e) {
@@ -225,7 +216,6 @@ export class MainComponent {
 
   mount() {
     this.mountPoint.innerHTML = this.render();
-    this.checkTheme();
     this.querySelectors();
     this.initMaterial();
     this.mountChildren();
