+++ src/index.js
@@ -5,5 +5,3 @@ const root = document.getElementById("root");
 
 const main = new MainComponent(root);
 main.mount();
-
-window.main = main;
