+++ src/index.scss
@@ -40,5 +40,3 @@
 @import "./components/SongsTable/SongsTable";
 @import "./components/LicenseDialog/LicenseDialog";
 @import "./components/DarkMode/DarkMode";
-
-@import "./components/DarkMode/DarkModeTheme";
